backend/services/quest-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose==3.3.0
pydantic==2.8.2
boto3==1.34.162   
pydantic_settings==2.10.1
python-multipart
pyjwt
httpx>=0.24,<1.0
